Move GlobalErrorAttributes and ErrorWebExceptionHandler in own package
